EmployeeApplication class added
