+++ src/test/java/com/knoldus/java/lambdaStream/EmployeeImplementationTest.java
@@ -28,7 +28,7 @@ class EmployeeImplementationTest {
     public void testSortByName(){
         List<Employee> employees = EmployeeImplementation.sortByName(employeeList);
         Assert.assertEquals("Employee{age=25, name='AASHIF', sureName='ALI'}",employees.get(0).toString());
-    }
+ }
 
     @Test
     public void testFilterAllEmployeeDataStartingNameWith_A(){
